# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The backend REST API for FraserTickets.",
        "title": "FraserTickets Backend",
        "contact": {
            "name": "Aritro Saha",
            "url": "http://www.aritrosaha.ca",
            "email": "aritro.saha729@gmail.com"
        },
        "version": "1.0"
    },
    "host": "frasertickets-backend.aritrosaha.ca",
    "basePath": "/",
    "paths": {
...
        "/events/{id}/tickets": {
            "get": {
                "description": "Get every ticket for an event. Only available to admins.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Get tickets for event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets": {
            "get": {
                "description": "List the tickets owned by the user sending the request. Available to all users.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "List the requesting user's tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create a new ticket. Only available to admins.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Create new ticket",
                "parameters": [
                    {
                        "description": "Ticket details",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ticketControllerCreateRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets/all": {
            "get": {
                "description": "List all tickets. Only available to admins.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "List all tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets/scan": {
            "post": {
                "description": "Scans in a ticket given the ticket ID. Only available to admins.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Scans a ticket",
                "parameters": [
                    {
                        "description": "Search query",
                        "name": "searchQuery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ticketControllerScanRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TicketScan"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets/search": {
            "post": {
                "description": "Search for a ticket by using the owner and associated event. Only available to admins.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Search for ticket using owner and event",
                "parameters": [
                    {
                        "description": "Search query",
                        "name": "searchQuery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ticketControllerSearchRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets/user/{uid}": {
            "get": {
                "description": "List the tickets owned by the user sending the request. Only available to admins.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "List a user's tickets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Get one ticket. Only available to admins and the ticket owner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Get one ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tickets/{id}/delete": {
            "post": {
                "description": "Deletes a ticket. Only available to admins.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "Delete a ticket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.ticketControllerCreateRequestBody": {
            "type": "object",
            "required": [
                "eventID",
                "studentNumber"
            ],
            "properties": {
                "eventID": {
                    "type": "string"
                },
                "studentNumber": {
                    "type": "string"
                }
            }
        },
        "controllers.ticketControllerScanRequestBody": {
            "type": "object",
            "required": [
                "ticketID"
            ],
            "properties": {
                "ticketID": {
                    "type": "string"
                }
            }
        },
        "controllers.ticketControllerSearchRequestBody": {
            "type": "object",
            "required": [
                "eventID",
                "ownerID"
            ],
            "properties": {
                "eventID": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "string"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_timestamp": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "location": {
                    "description": "Ex. name of venue",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_timestamp": {
                    "type": "string"
                }
            }
        },
        "models.Ticket": {
            "type": "object",
            "properties": {
                "eventData": {
                    "$ref": "#/definitions/models.Event"
                },
                "eventID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastScanTime": {
                    "type": "string"
                },
                "ownerData": {
                    "$ref": "#/definitions/models.User"
                },
                "ownerID": {
                    "description": "owner ID",
                    "type": "string"
                },
                "scanCount": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.TicketScan": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "ticketData": {
                    "$ref": "#/definitions/models.Ticket"
                },
                "timestamp": {
                    "type": "string"
                },
                "userData": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "description": "This is also the UUID in Firebase Auth",
                    "type": "string"
                },
                "pfp_url": {
                    "type": "string"
                },
                "student_number": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyHeader": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}